Restructure signaler, fix race condition
